+++ ddxApplicationSettings_by_sKeyword/ddxAdobeAcrobatReaderSettings_by_sKeyword.py
@@ -23,8 +23,10 @@ ddxAdobeAcrobatReaderSettings_by_sKeyword = {
     "fasGetOptionalArguments": fasGetOptionalArguments,
     "dxConfigSettings": dxConfigSettings,
     # Adobe Reader has a component that constantly crashes with a NULL pointer when you enable page heap.
-    "asBinaryNamesThatAreAllowedToRunWithoutPageHeap": ["rdrcef.exe"],
-
+    "asBinaryNamesThatAreAllowedToRunWithoutPageHeap": [
+      "AcroCEF.exe",
+      "rdrcef.exe",
+    ],
   },
   "acrobat_x86": {
     "sBinaryPath": sApplicationBinaryPath_x86,
@@ -32,6 +34,9 @@ ddxAdobeAcrobatReaderSettings_by_sKeyword = {
     "dxConfigSettings": dxConfigSettings,
     "sISA": "x86",
     # Adobe Reader has a component that constantly crashes with a NULL pointer when you enable page heap.
-    "asBinaryNamesThatAreAllowedToRunWithoutPageHeap": ["rdrcef.exe"],
+    "asBinaryNamesThatAreAllowedToRunWithoutPageHeap": [
+      "AcroCEF.exe",
+      "rdrcef.exe"
+    ],
   },
 };