+++ ddxApplicationSettings_by_sKeyword/ddxMozillaFirefoxSettings_by_sKeyword.py
@@ -1,7 +1,7 @@
 import os;
 from dxConfig import dxConfig;
 from fsFirstExistingFile import fsFirstExistingFile;
-from mFileSystem import mFileSystem;
+import mFileSystem2;
 sProgramFilesPath_x86 = os.getenv("ProgramFiles(x86)") or os.getenv("ProgramFiles");
 sProgramFilesPath_x64 = os.getenv("ProgramW6432");
 
@@ -18,27 +18,14 @@ sApplicationBinaryPath_x86 = fsFirstExistingFile(
 );
 sApplicationBinaryPath = sApplicationBinaryPath_x64 or sApplicationBinaryPath_x86;
 
-sFirefoxProfilePath = mFileSystem.fsPath(os.getenv("TEMP"), "Firefox-profile");
+oTempFolder = mFileSystem2.foGetFolder(os.getenv("TEMP"));
+oFirefoxProfileFolder = oTempFolder.foGetOrCreateChildFolder("Firefox-profile");
 
 def fasGetFirefoxStaticArguments(bForHelp):
-  if bForHelp:
-    # The folder may not exist at this point, so we cannot guarantee a 8.3 path
-    # exists. Also, the 8.3 path may not be easily readable. Therefore, we'll
-    # always use the long path in the help.
-    sUsedFirefoxProfilePath = sFirefoxProfilePath;
-  else:
-    # Firefox cannot handle long paths (starting with "\\?\") so we'll use the
-    # 8.3 path to make sure it will work. To get an 8.3 path, there should be a
-    # file or folder for that path. In this case, we want a folder, so we'll
-    # make sure it's created if it does not exist yet.
-    if not mFileSystem.fbIsFolder(sFirefoxProfilePath):
-      assert mFileSystem.fbCreateFolder(sFirefoxProfilePath), \
-          "Cannot create Firefox profile folder %s" % sFirefoxProfilePath;
-    sUsedFirefoxProfilePath = mFileSystem.fs83Path(sFirefoxProfilePath)
+  oFirefoxProfileFolder.fCreate();
   return [
     "--no-remote",
-    "-profile",
-        sUsedFirefoxProfilePath,
+    "-profile", oFirefoxProfileFolder.sPath,
   ];
 
 def fasGetFirefoxOptionalArguments(bForHelp = False):
@@ -51,20 +38,15 @@ def fFirefoxSetup(bFirstRun):
     os.environ["MOZ_DISABLE_GMP_SANDBOX"] = "1";
     os.environ["MOZ_DISABLE_NPAPI_SANDBOX"] = "1";
     os.environ["MOZ_DISABLE_GPU_SANDBOX "] = "1";
-  # Delete the profile before the application runs, so as to start with a clean state, and not to keep state between
-  # different runs of the application.
-  fDeleteProfile();
+  # We want to start with a clean state and we use an empty profile folder to
+  # do that; create the profile folder if it does not exist and delete everything
+  # in it if it does exist.
+  oFirefoxProfileFolder.fCreate();
+  oFirefoxProfileFolder.fDeleteChildren();
 
 def fFirefoxCleanup():
   # Delete the profile to clean up after application ran.
-  fDeleteProfile();
-
-def fDeleteProfile():
-  if mFileSystem.fbIsFolder(sFirefoxProfilePath):
-    mFileSystem.fbDeleteChildrenFromFolder(sFirefoxProfilePath);
-  else:
-    assert mFileSystem.fbCreateFolder(sFirefoxProfilePath), \
-        "Cannot create Firefox profile folder %s" % sFirefoxProfilePath;
+  oFirefoxProfileFolder.fDeleteChildren();
 
 # Known applications can have regular expressions that map source file paths in its output to URLs, so the details HTML for any detected bug can have clickable
 # links to an online source repository: