+++ ddxApplicationSettings_by_sKeyword/fsFirstExistingFile.py
@@ -1,7 +1,7 @@
-import mFileSystem;
+import mFileSystem2;
 def fsFirstExistingFile(*asPossiblePaths):
   for sPossiblePath in asPossiblePaths:
-    if sPossiblePath and mFileSystem.fbIsFile(sPossiblePath):
+    if sPossiblePath and mFileSystem2.fo0GetFile(sPossiblePath):
       return sPossiblePath;
   return None;