+++ RELEASE NOTES.md
@@ -1,3 +1,10 @@
+2019-01-09
+==========
+
+Internal changes
+----------------
++ Switched from mFileSystem to mFileSystem2
+
 2018-11-21
 ==========
 Externally noticeable changes